+++ include/libpkgtransaction/error.h
@@ -24,6 +24,7 @@ enum class error_code {
 class error : public std::runtime_error {
 public:
   error(error_code code, std::string message);
+  ~error() override;
   [[nodiscard]] error_code code() const noexcept;
 private:
   error_code code_;